+++ app/components/views/ViewMap.vue
@@ -967,8 +967,8 @@ await setupDataFetch({
                     collapsed: false,
                 }),
             ],
+            maxTilesLoading: 1000,
             interactions: [],
-            maxTilesLoading: 128,
             view: new View({
                 center,
                 zoom,